Extract useNodeAttributionResolver hook
Move the SourceInfoReconstructor + getNodeAttribution cache logic out of
AstRenderer and into hooks/useAttribution.ts, alongside AttributionContext.
NodeAttributionContext moves too, since nothing about its shape is
debug-specific. Any renderer with an astContext can now opt in with one
hook call and a provider wrap instead of re-deriving the plumbing.
